# docs/swagger.yaml
basePath: /api/v1
definitions:
  common.ResponseMidtrans:
    properties:
      redirect_url:
        type: string
      token:
        type: string
    type: object
  dto.AdminRequestDto:
    properties:
      email:
        type: string
      name:
        type: string
      password:
        type: string
      username:
        type: string
    type: object
  dto.LoginRequestDto:
    properties:
      password:
        type: string
      username:
        type: string
    required:
    - password
    - username
    type: object
  dto.LoginResponseDto:
    properties:
      accessToken:
        type: string
      userId:
        type: string
    type: object
...
  dto.TopupRequest:
    properties:
      ammount:
        type: integer
    type: object
...
  model.Admin:
    properties:
      created_at:
        type: string
      email:
        type: string
      id:
        type: string
      name:
        type: string
      password:
        type: string
      role:
        type: string
      updated_at:
        type: string
      username:
        type: string
    type: object
...
  model.TableTopupPayment:
    properties:
      ammount:
        type: integer
      created_at:
        type: string
      deskripsi:
        type: string
      id:
        type: string
      status:
        type: string
      token_midtrans:
        type: string
      updated_at:
        type: string
      url_payment:
        type: string
      user_id:
        type: string
    type: object
...
  model.User:
    properties:
      created_at:
        type: string
      email:
        type: string
      id:
        type: string
      name:
        type: string
      password:
        type: string
      phone_number:
        type: string
      role:
        type: string
      saldo:
        type: integer
      updated_at:
        type: string
      username:
        type: string
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: A tag service API in Go using Gin framework
  title: Tag Service API
  version: "1.0"
paths:
  /admin:
    post:
      consumes:
      - application/json
      description: Admin can create account
      parameters:
      - description: payload register
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/dto.AdminRequestDto'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/model.Admin'
      summary: Admin can create account
      tags:
      - Admin
  /admin/login:
    post:
      consumes:
      - application/json
      description: Admin can login
      parameters:
      - description: payload login
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/dto.LoginRequestDto'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dto.LoginResponseDto'
      summary: Admin can login
      tags:
      - Admin
  /admin/user/{id}:
    get:
      consumes:
      - application/json
      description: Admin can get info and balance user
      parameters:
      - description: Bearer
        in: header
        name: Authorization
        required: true
        type: string
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/model.User'
      security:
      - ApiKeyAuth: []
      summary: Admin can get info and balance user
      tags:
      - Admin
  /topup:
    post:
      consumes:
      - application/json
      description: User can topup with payment gateway
      parameters:
      - description: Bearer
        in: header
        name: Authorization
        required: true
        type: string
      - description: Ammount topup
        in: body
        name: ammount
        required: true
        schema:
          $ref: '#/definitions/dto.TopupRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/common.ResponseMidtrans'
      security:
      - ApiKeyAuth: []
      summary: User can topup with payment gateway
      tags:
      - Topup
  /topup/history:
    get:
      consumes:
      - application/json
      description: User can get histories topup with payment gateway
      parameters:
      - description: Bearer
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.TableTopupPayment'
            type: array
      security:
      - ApiKeyAuth: []
      summary: User can get histories topup with payment gateway
      tags:
      - Topup
  /topup/history/{id}:
    get:
      consumes:
      - application/json
      description: Admin can get user histories topup with payment gateway
      parameters:
      - description: Bearer
        in: header
        name: Authorization
        required: true
        type: string
      - description: User id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.TableTopupPayment'
            type: array
      security:
      - ApiKeyAuth: []
      summary: Admin can get user histories topup with payment gateway
      tags:
      - Admin
...
swagger: "2.0"